added unit tests
